fix: unify Shepherd tours — single entry point via profile menu
- Remove usePageTour auto-launch (caused surprise popups on every page)
- Fix profile menu "Tour this page" to use getTourForRoute() registry
  instead of always launching welcome tour
- Fix "Reset All Tours" to use resetAllTours() (was using resetOnboarding)
- Delete dead GuideButton FAB (never rendered anywhere)
- Delete orphaned workflows-tour.ts (never registered)
- Delete tour-bridge.ts + use-tour-tab-bridge.ts (overengineered, unused)
- Remove useTourTabBridge from create-agent-modal

Tours now: one welcome + 8 page-specific, all via profile menu only.
