+++ frontend/components/auth/profile-menu.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react'
 import { useRouter } from 'next/navigation'
 import { useUser, useClerk } from '@clerk/nextjs'
-import { User, Settings, LogOut, ChevronDown, Sparkles, Map, RotateCcw, ChevronRight } from 'lucide-react'
+import { User, Settings, LogOut, ChevronDown, Sparkles, Compass, RotateCcw, ChevronRight } from 'lucide-react'
 import * as DropdownMenu from '@radix-ui/react-dropdown-menu'
 import { usePathname } from 'next/navigation'
 
@@ -149,28 +149,34 @@ export function ProfileMenu() {
                                         setOpen(false)
                                         setTimeout(async () => {
                                             if (user) {
-                                                const { createWelcomeTour } = await import('@/lib/shepherd/tours/welcome-tour')
-                                                const tour = createWelcomeTour(user.id)
+                                                const { createWelcomeTour } = await import('@/lib/shepherd/tour-registry')
+                                                const { resetTour } = await import('@/lib/shepherd/tour-storage')
+                                                resetTour('welcome', user.id)
+                                                const tour = await createWelcomeTour(user.id)
                                                 tour.start()
                                             }
                                         }, 150)
                                     }}
                                 >
-                                    <Map className="w-4 h-4 text-orange-400" />
+                                    <Compass className="w-4 h-4 text-orange-400" />
                                     <span>Welcome Orientation</span>
                                 </DropdownMenu.Item>
 
                                 <DropdownMenu.Item
                                     className="flex items-center space-x-3 px-3 py-2.5 rounded-lg text-sm text-slate-300 hover:text-white hover:bg-slate-800/80 cursor-pointer outline-none transition-colors"
+                                    disabled={!pathname}
                                     onSelect={() => {
                                         setOpen(false)
                                         setTimeout(async () => {
-                                            if (user) {
-                                                // Page-specific tours can be added later;
-                                                // for now re-launch the main tour
-                                                const { createWelcomeTour } = await import('@/lib/shepherd/tours/welcome-tour')
-                                                const tour = createWelcomeTour(user.id)
-                                                tour.start()
+                                            if (user && pathname) {
+                                                const { getTourForRoute } = await import('@/lib/shepherd/tour-registry')
+                                                const { resetTour } = await import('@/lib/shepherd/tour-storage')
+                                                const entry = getTourForRoute(pathname)
+                                                if (entry) {
+                                                    resetTour(entry.id, user.id)
+                                                    const tour = await entry.factory(user.id)
+                                                    tour.start()
+                                                }
                                             }
                                         }, 150)
                                     }}
@@ -186,8 +192,8 @@ export function ProfileMenu() {
                                     className="flex items-center space-x-3 px-3 py-2.5 rounded-lg text-sm text-slate-400 hover:text-slate-200 hover:bg-slate-800/80 cursor-pointer outline-none transition-colors"
                                     onSelect={() => {
                                         if (user) {
-                                            import('@/lib/shepherd/tour-storage').then(({ resetOnboarding }) => {
-                                                resetOnboarding(user.id)
+                                            import('@/lib/shepherd/tour-storage').then(({ resetAllTours }) => {
+                                                resetAllTours(user.id)
                                             })
                                         }
                                         setOpen(false)
